generalized getPaddedNumber()

+++ Trivia/Responses/JsonResponseSerializer.cpp
@@ -72,7 +72,7 @@ buffer JsonResponseSerializer::serializeGeneralResponse(const messageType type, 
     buff.append_range(std::to_string(type));
 
     // Pushing the JSON's length to the buffer
-    buff.append_range(Helper::getPaddedNumber(static_cast<uint32_t>(json.length()), BYTES_RESERVED_FOR_MSG_LEN));
+    buff.append_range(Helper::getPaddedNumber(json.length(), BYTES_RESERVED_FOR_MSG_LEN));
 
     // Pushing the actual message to the buffer
     buff.append_range(json);
